pjt-back/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
